perf(cli): async handler (function, job, layout, model, ogImage) (ESM prep)
