+++ packages/cli/src/commands/destroy/layout/__tests__/layout.test.js
@@ -14,7 +14,7 @@ import { vi, beforeEach, afterEach, test, expect } from 'vitest'
 
 import '../../../../lib/test'
 
-import { files } from '../../../generate/layout/layout.js'
+import { files } from '../../../generate/layout/layoutHandler.js'
 import { tasks } from '../layout.js'
 
 beforeEach(() => {
